move the ui related files to ui package
