--- app/src/main/kotlin/ir/miare/androidcodechallenge/theme/Type.kt
@@ -1,5 +1,0 @@
-package ir.miare.androidcodechallenge.theme
-
-import androidx.compose.material3.Typography
-
-val AppTypography = Typography()
