+++ app/src/main/kotlin/ir/miare/androidcodechallenge/ui/theme/Type.kt
@@ -1,0 +1,5 @@
+package ir.miare.androidcodechallenge.ui.theme
+
+import androidx.compose.material3.Typography
+
+internal val AppTypography = Typography()
